+++ types/index.d.ts
@@ -1,8 +1,10 @@
-import MarkdownIt = require('markdown-it');
-import Token = require('markdown-it/lib/token');
-import State = require('markdown-it/lib/rules_core/state_core');
+import MarkdownIt from 'markdown-it';
+import { default as MarkdownItToken } from 'markdown-it/lib/token.mjs';
+import { default as MarkdownItState} from 'markdown-it/lib/rules_core/state_core.mjs';
 
 declare namespace anchor {
+  export type Token = MarkdownItToken
+  export type State = MarkdownItState
   export type RenderHref = (slug: string, state: State) => string;
   export type RenderAttrs = (slug: string, state: State) => Record<string, string | number>;
 
@@ -13,26 +15,31 @@ declare namespace anchor {
     renderAttrs?: RenderAttrs
   }
 
+  export interface HeaderLinkPermalinkOptions extends PermalinkOptions {
+    safariReaderFix?: boolean;
+  }
+
   export interface LinkAfterHeaderPermalinkOptions extends PermalinkOptions {
     style?: 'visually-hidden' | 'aria-label' | 'aria-describedby' | 'aria-labelledby',
     assistiveText?: (title: string) => string,
     visuallyHiddenClass?: string,
-    space?: boolean,
+    space?: boolean | string,
     placement?: 'before' | 'after'
+    wrapper?: [string, string] | null
   }
 
   export interface LinkInsideHeaderPermalinkOptions extends PermalinkOptions {
-    space?: boolean,
+    space?: boolean | string,
     placement?: 'before' | 'after',
     ariaHidden?: boolean
   }
 
   export interface AriaHiddenPermalinkOptions extends PermalinkOptions {
-    space?: boolean,
+    space?: boolean | string,
     placement?: 'before' | 'after'
   }
 
-  export type PermalinkGenerator = (slug: string, opts: PermalinkOptions, state: State, index: number) => string;
+  export type PermalinkGenerator = (slug: string, opts: PermalinkOptions, state: State, index: number) => void;
 
   export interface AnchorInfo {
     slug: string;
@@ -43,6 +50,8 @@ declare namespace anchor {
     level?: number | number[];
 
     slugify?(str: string): string;
+    slugifyWithState?(str: string, state: State): string;
+    getTokensText?(tokens: Token[]): string;
 
     uniqueSlugStartIndex?: number;
     permalink?: PermalinkGenerator;
@@ -53,10 +62,10 @@ declare namespace anchor {
   }
 
   export const permalink: {
-    headerLink: () => PermalinkGenerator
-    linkAfterHeader: (opts: LinkAfterHeaderPermalinkOptions) => PermalinkGenerator
-    linkInsideHeader: (opts: LinkInsideHeaderPermalinkOptions) => PermalinkGenerator
-    ariaHidden: (opts: AriaHiddenPermalinkOptions) => PermalinkGenerator
+    headerLink: (opts?: HeaderLinkPermalinkOptions) => PermalinkGenerator
+    linkAfterHeader: (opts?: LinkAfterHeaderPermalinkOptions) => PermalinkGenerator
+    linkInsideHeader: (opts?: LinkInsideHeaderPermalinkOptions) => PermalinkGenerator
+    ariaHidden: (opts?: AriaHiddenPermalinkOptions) => PermalinkGenerator
   };
 }